+++ CHANGELOG.md
@@ -18,6 +18,7 @@ accidentally triggering the load of a previous DB version.**
 * #3801 Fail size utility functions when data nodes do not respond
 * #3781 Segfault in fill_result_error()
 * #3779 Support GRANT/REVOKE on distributed database
+* #3808 Properly handle `max_retries` option
 
 **Thanks**
 * @cbisnett for reporting and fixing a typo in an error message